requirements.txt
faiss-cpu
fastapi
lxml
numpy
openai
pandas
pillow
pymupdf
pypdf
python-dotenv
pytesseract
requests
sentence-transformers
streamlit
torch
transformers
tqdm
uvicorn